server/requirements.txt
PyYAML~=5.1.2
netaddr~=0.8.0
asyncio~=3.4.3
aiohttp~=3.6.2
asfpy~=0.22
multipart~=0.2.1
aiosmtplib~=1.1.3
netaddr~=0.8.0
aiofiles
